update gerer controler signer

+++ app/Http/Controllers/gerercontroller.php
@@ -2,14 +2,10 @@
 
 namespace App\Http\Controllers;
 
-
 use App\Http\Requests\gererRequest;
-use App\Models\administrateur;
 use App\Models\gerer;
-use App\Models\session;
 use Exception;
-use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\DB;
 
 class gererController extends Controller
 {
@@ -40,13 +36,11 @@ class gererController extends Controller
     
         public function update(gererRequest $request,  $id_administrateur,$id_session) {
             try {
-                $gerer = Gerer::where('id_admin', $id_administrateur)
-                      ->where('id_session', $id_session)
-                      ->firstOrFail();
-
-                // Mettre à jour les données
-                $gerer->update($request->all());
-        
+                
+                $gerer=DB::table('gerers')->where('id_administrateur',$id_administrateur)
+                ->where('id_session',$id_session)
+                ->update(['id_administrateur'=>$request->id_administrateur,'id_session'=>$request->id_session]);
+                
                 return response()->json([
                     'status_code' => 201,
                     'status_message' => 'La gestion a été modifiée',
@@ -57,9 +51,9 @@ class gererController extends Controller
                 return response()->json($exception);
             }
         }
-            public function delete(gerer $gerer) {
+            public function delete(  $id_administrateur,$id_session) {
                 try{
-                       $gerer->delete();
+                    $gerer = DB::table('gerers')->where('id_administrateur',$id_administrateur)->where('id_session',$id_session)->delete();
        
                    return response()->json([
                        'status_code'=>200,
